une modification de la gestion de base de données ainsi faire des tests et fixer les erreurs avant la soutenance

+++ routes/web.php
@@ -19,11 +19,10 @@ use App\Http\Controllers\EmailController;
 use Illuminate\Support\Facades\Auth;
 
 
-
+//->middleware('auth')
 Route::get('/', function () {
     return view('welcome');
 });
-
 Route::get('/homepage', [homepageController::class, 'index'])->name('homepage.index');
 Route::post('/homepage', [homepageController::class, 'store'])->name('homepage.store');
 Route::get('/loginpage', [loginpageController::class, 'login'])->name('loginpage.login');////////
@@ -40,79 +39,79 @@ Route::get('/profilepage', [profilepageController::class, 'index'])->name('profi
 
 
 Route::get('/profilepage/edit', [profilepageController::class, 'edit'])->name('profilepage.edit')->middleware('auth');
-Route::put('/profilepage', [profilepageController::class, 'update'])->name('profilepage.update');
+Route::put('/profilepage', [profilepageController::class, 'update'])->name('profilepage.update')->middleware('auth');
 
 Route::get('/taskmanguser', [taskmanguserController::class, 'index'])->name('taskmanguser.index')->middleware('auth');
 Route::get('/taskmangchef', [taskmangchefController::class, 'index'])->name('taskmangchef.index')->middleware('auth');
 
 Route::get('/taskmangchef/create', [taskmangchefController::class, 'create'])->name('taskmangchef.create')->middleware('auth');
-Route::post('/taskmangchef', [taskmangchefController::class, 'store'])->name('taskmangchef.store');
+Route::post('/taskmangchef', [taskmangchefController::class, 'store'])->name('taskmangchef.store')->middleware('auth');
 Route::get('/taskmanguser/{task}', [taskmanguserController::class,'show'])->name('taskmanguser.show')->middleware('auth');
 Route::get('/taskmangchef/{task}', [taskmangchefController::class,'show'])->name('taskmangchef.show')->middleware('auth');
 Route::get('/taskmangchef/{task}/edit', [taskmangchefController::class, 'edit'])->name('taskmangchef.edit')->middleware('auth');
-route::put('/taskmangchef/{task}', [taskmangchefController::class, 'update'])->name('taskmangchef.update');
+route::put('/taskmangchef/{task}', [taskmangchefController::class, 'update'])->name('taskmangchef.update')->middleware('auth');
 Route::get('/breakrequest', [breakrequestController::class, 'index'])->name('breakrequest.index')->middleware('auth');
-route::post('/breakrequest', [breakrequestController::class, 'store'])->name('breakrequest.store');
+route::post('/breakrequest', [breakrequestController::class, 'store'])->name('breakrequest.store')->middleware('auth');
 Route::get('/breakrequest/create', [breakrequestController::class, 'create'])->name('breakrequest.create')->middleware('auth');
 route::get('/breakrequestadmin/{breakrequestadmin}', [breakrequestadminController::class, 'show'])->name('breakrequestadmin.show')->middleware('auth');
 Route::get('/breakrequestadmin', [breakrequestadminController::class, 'index'])->name('breakrequestadmin.index')->middleware('auth');
-route::post('/breakrequestadmin/{breakrequestadmin}/accept', [breakrequestadminController::class, 'accept'])->name('breakrequestadmin.accept');
-route::post('/breakrequestadmin/{breakrequestadmin}/refuse', [breakrequestadminController::class, 'refuse'])->name('breakrequestadmin.refuse');
+route::post('/breakrequestadmin/{breakrequestadmin}/accept', [breakrequestadminController::class, 'accept'])->name('breakrequestadmin.accept')->middleware('auth');
+route::post('/breakrequestadmin/{breakrequestadmin}/refuse', [breakrequestadminController::class, 'refuse'])->name('breakrequestadmin.refuse')->middleware('auth');
 
 
-Route::get('/contactmssg', [contactmssgController::class, 'index'])->name('contactmssg.index')->middleware('auth');
+Route::get('/contactmssg', [contactmssgController::class, 'index'])->name('contactmssg.index')->middleware('auth')->middleware('auth');
 Route::get('/contactmssg/{contactmssgId}/edit', [contactmssgController::class, 'edit'])->name('contactmssg.edit')->middleware('auth');
-route::put('/contactmssg/{contactmssgId}', [contactmssgController::class, 'update'])->name('contactmssg.update');
+route::put('/contactmssg/{contactmssgId}', [contactmssgController::class, 'update'])->name('contactmssg.update')->middleware('auth');
 route::get('/contactmssg/{contactmssgId}', [contactmssgController::class, 'show'])->name('contactmssg.show')->middleware('auth');
 
 
 
 Route::get('/signuprequest', [signuprequestController::class, 'index'])->name('signuprequest.index')->middleware('auth');
 route::get('/signuprequest/{signuprequestId}', [signuprequestController::class,'show'])->name('signuprequest.show')->middleware('auth');
-Route::delete('/signuprequest/{signuprequestId}', [signuprequestController::class,'refuse'])->name('signuprequest.refuse');
-Route::post('/signuprequest/{signuprequestId}', [signuprequestController::class,'accept'])->name('signuprequest.accept');
+Route::delete('/signuprequest/{signuprequestId}', [signuprequestController::class,'refuse'])->name('signuprequest.refuse')->middleware('auth');
+Route::post('/signuprequest/{signuprequestId}', [signuprequestController::class,'accept'])->name('signuprequest.accept')->middleware('auth');
 
 Route::get('/usersprofiles', [usersprofilesController::class, 'index'])->name('usersprofiles.index')->middleware('auth');//////////:///////
 route::get('/usersprofiles/create', [usersprofilesController::class, 'create'])->name('usersprofiles.create')->middleware('auth');
 route::get('/usersprofiles/{usersprofile}/edit', [usersprofilesController::class, 'edit'])->name('usersprofiles.edit')->middleware('auth');
-route::put('/usersprofiles/{usersprofile}', [usersprofilesController::class, 'update'])->name('usersprofiles.update');
-route::post('/usersprofiles', [usersprofilesController::class, 'store'])->name('usersprofiles.store');
+route::put('/usersprofiles/{usersprofile}', [usersprofilesController::class, 'update'])->name('usersprofiles.update')->middleware('auth');
+route::post('/usersprofiles', [usersprofilesController::class, 'store'])->name('usersprofiles.store')->middleware('auth');
 route::get('/usersprofiles/{usersprofile}', [usersprofilesController::class,'show'])->name('usersprofiles.show')->middleware('auth');
 
 Route::get('/dashboard', [dashboardController::class, 'index'])->name('dashboard.index')->middleware('auth');
 
 Route::get('/dashboard/create', [dashboardController::class, 'create'])->name('dashboard.create')->middleware('auth');
-route::post('/dashboard', [dashboardController::class, 'store'])->name('dashboard.store');
+route::post('/dashboard', [dashboardController::class, 'store'])->name('dashboard.store')->middleware('auth');
 
 
 Route::get('/dashboard/{task}/edit', [dashboardController::class, 'edit'])->name('dashboard.edit')->middleware('auth');
-route::put('/dashboard/{task}', [dashboardController::class, 'update'])->name('dashboard.update');
+route::put('/dashboard/{task}', [dashboardController::class, 'update'])->name('dashboard.update')->middleware('auth');
 
-route::delete('/dashboard/{task}', [dashboardController::class,'destroy'])->name('dashboard.destroy');
+route::delete('/dashboard/{task}', [dashboardController::class,'destroy'])->name('dashboard.destroy')->middleware('auth');
 
 
 Route::get('/chef_dashboard', [chef_dashboardController::class, 'index'])->name('chef_dashboard.index')->middleware('auth');///////://////
 Route::get('/chef_dashboard/create', [chef_dashboardController::class, 'create'])->name('chef_dashboard.create')->middleware('auth');
-route::post('/chef_dashboard', [chef_dashboardController::class,'store'])->name('chef_dashboard.store');
+route::post('/chef_dashboard', [chef_dashboardController::class,'store'])->name('chef_dashboard.store')->middleware('auth');
 Route::get('/chef_dashboard/{task}/edit', [chef_dashboardController::class, 'edit'])->name('chef_dashboard.edit')->middleware('auth');
-route::put('/chef_dashboard/{task}', [chef_dashboardController::class, 'update'])->name('chef_dashboard.update');
-route::delete('/chef_dashboard/{task}', [chef_dashboardController::class,'destroy'])->name('chef_dashboard.destroy');
+route::put('/chef_dashboard/{task}', [chef_dashboardController::class, 'update'])->name('chef_dashboard.update')->middleware('auth');
+route::delete('/chef_dashboard/{task}', [chef_dashboardController::class,'destroy'])->name('chef_dashboard.destroy')->middleware('auth');
 route::get('/chef_breakrequest', [chef_breakrequestController::class, 'index'])->name('chef_breakrequest.index')->middleware('auth');
 route::get('/chef_breakrequest/create', [chef_breakrequestController::class, 'create'])->name('chef_breakrequest.create')->middleware('auth');
-route::post('/chef_breakrequest', [chef_breakrequestController::class,'store'])->name('chef_breakrequest.store');
+route::post('/chef_breakrequest', [chef_breakrequestController::class,'store'])->name('chef_breakrequest.store')->middleware('auth');
 Route::get('/chef_profile', [chef_profileController::class, 'index'])->name('chef_profile.index')->middleware('auth');
 Route::get('/chef_profile/edit', [chef_profileController::class, 'edit'])->name('chef_profile.edit')->middleware('auth');
-Route::put('/chef_profile', [chef_profileController::class, 'update'])->name('chef_profile.update');
+Route::put('/chef_profile', [chef_profileController::class, 'update'])->name('chef_profile.update')->middleware('auth');
 
-Route::get('send_request-accepted_email/{signuprequestId}' , [EmailController::class, 'Send_RequestAccepted_Email'])->name('send_request-accepted_email');
-route::get('send_request-refused_email/{signuprequestId}', [EmailController::class, 'Send_RequestRefused_Email'])->name('send_request-refused_email');
-Route::get('send_contactmssgResponse_email/{contactmssgId}', [EmailController::class, 'send_contactmssgResponse_Email'])->name('send_contactmssgResponse_email');
+Route::get('send_request-accepted_email/{signuprequestId}' , [EmailController::class, 'Send_RequestAccepted_Email'])->name('send_request-accepted_email')->middleware('auth');
+route::get('send_request-refused_email/{signuprequestId}', [EmailController::class, 'Send_RequestRefused_Email'])->name('send_request-refused_email')->middleware('auth');
+Route::get('send_contactmssgResponse_email/{contactmssgId}', [EmailController::class, 'send_contactmssgResponse_Email'])->name('send_contactmssgResponse_email')->middleware('auth');
 
-Route::get('/dashboard/{id}/state', [dashboardController::class, 'state'])->name('dashboard.state');
+Route::get('/dashboard/{id}/state', [dashboardController::class, 'state'])->name('dashboard.state')->middleware('auth');
 route::get('/dashboard/{id}/show', [dashboardController::class, 'show'])->name('dashboard.show')->middleware('auth');
 
-Route::get('/chef_dashboard/{id}/state', [chef_dashboardController::class, 'state'])->name('chef_dashboard.state');
+Route::get('/chef_dashboard/{id}/state', [chef_dashboardController::class, 'state'])->name('chef_dashboard.state')->middleware('auth');
 route::get('/chef_dashboard/{id}/show', [chef_dashboardController::class, 'show'])->name('chef_dashboard.show')->middleware('auth');
 
-Route::get('/taskmanguser/{id}/state', [taskmanguserController::class, 'state'])->name('taskmanguser.state');
+Route::get('/taskmanguser/{id}/state', [taskmanguserController::class, 'state'])->name('taskmanguser.state')->middleware('auth');
 Route::delete('/usersprofiles/{userId}', [usersprofilesController::class, 'destroy'])->name('usersprofiles.destroy')->middleware('auth');
